force all values to double

+++ src/fortran_routing/mc_pylink_v00/MC_singleSeg_singleTS/MCsingleSegStime_f2py_noIC_clean.f90
@@ -29,8 +29,8 @@ contains
 
         integer :: iter
         integer :: maxiter, tries
-        real*8 :: mindepth, aerror, rerror
-        real*8 :: R, Twl, h_1 , h, h_0, Qj, Qj_0
+        real(dp) :: mindepth, aerror, rerror
+        real(dp) :: R, Twl, h_1 , h, h_0, Qj, Qj_0
 
         !* parameters of Secant method
         maxiter  = 100
@@ -156,9 +156,9 @@ contains
 
         implicit none
 
-        real*8, intent(in) :: h_0  !, refQj_0
-        real*8, intent(out) :: Qj_0
-        real*8 :: Twl, AREA, WP, R, Ck, Km, X, D
+        real(dp), intent(in) :: h_0  !, refQj_0
+        real(dp), intent(out) :: Qj_0
+        real(dp) :: Twl, AREA, WP, R, Ck, Km, X, D
 
         !**top surface water width of the channel inflow
         Twl = Bw + 2.0*z*h_0
@@ -244,9 +244,9 @@ contains
 
         implicit none
 
-        real*8,intent(in) :: h
-        real*8,intent(out) :: Qj
-        real*8 :: Twl, AREA, WP, R, Ck, Km, X, D
+        real(dp),intent(in) :: h
+        real(dp),intent(out) :: Qj
+        real(dp) :: Twl, AREA, WP, R, Ck, Km, X, D
 
         !--upper interval -----------
         Twl = Bw + 2.0*z*h  !--top width of the channel inflow
